feat(starter_1s): assemblyStruct for extracting elements
