+++ Project1/starter_1s/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "files.eol": "\r\n"
+}
