+++ Project1/starter_1s/.vscode/tasks.json
@@ -1,0 +1,20 @@
+{
+    "tasks": [{
+        "type": "cppbuild",
+        "label": "C/C++: make simulator",
+        "command": "/usr/bin/make",
+        "args": ["assembler"],
+        "options": {
+            "cwd": "${workspaceFolder}"
+        },
+        "problemMatcher": ["$gcc"],
+        "group": {
+            "kind": "build",
+            "isDefault": true
+        },
+        "detail": "Task generated by Debugger."
+    }],
+    "version": "2.0.0"
+}
+
+
