+++ Project1/starter_1s/checking_test/size_t_check.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+int main() {
+    int s = 10;
+    printf("Size: %zu Bytes\n", sizeof(s));
+    return 0;
+}
